Fixes an issue with block changes (and 16-bit block ids), as well as some other instances were byte was still used for block ids (only testcases).  #1210

+++ engine-tests/src/test/java/org/terasology/world/propagation/BetweenChunkPropagationTest.java
@@ -81,7 +81,7 @@ public class BetweenChunkPropagationTest extends TerasologyTestingEnvironment {
         solid = new Block();
         solid.setDisplayName("Solid");
         solid.setUri(new BlockUri("engine:solid"));
-        solid.setId((byte) 5);
+        solid.setId((short) 5);
         for (Side side : Side.values()) {
             solid.setFullSide(side, true);
         }
