+++ engine-tests/src/test/java/org/terasology/world/propagation/BulkSunlightPropagationTest.java
@@ -69,7 +69,7 @@ public class BulkSunlightPropagationTest extends TerasologyTestingEnvironment {
         solid = new Block();
         solid.setDisplayName("Solid");
         solid.setUri(new BlockUri("engine:solid"));
-        solid.setId((byte) 5);
+        solid.setId((short) 5);
         for (Side side : Side.values()) {
             solid.setFullSide(side, true);
         }
